pyproject.toml
[project]
name = "cioos-metadata-conversion"
version = "0.1.0"
description = ""
authors = [{ name = "Jessy Barrette", email = "30420025+JessyBarrette@users.noreply.github.com" }]
requires-python = "~=3.11"
readme = "README.md"
dependencies = [
    "loguru>=0.7.2,<0.8",
    "requests>=2.32.3,<3",
    "cffconvert>=2.0.0,<3",
    "pycountry>=24.6.1,<25",
    "pyyaml>=6.0.1,<7",
    "lxml>=5.2.2,<6",
    "metadata-xml",
    "click>=7,<9",
    "datacite>=1.2.0",
    "firebase-to-xml",
    "python-dotenv>=1.1.0",

]

[project.scripts]
cioos_metadata_conversion = "cioos_metadata_conversion.__main__:cli"

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.2,<9",
    "ruff>=0.5.0,<0.6",
    "pytest-xdist>=3.6.1,<4",
]

[tool.uv.sources]
metadata-xml = { git = "https://github.com/cioos-siooc/metadata-xml.git" }
firebase-to-xml = { git = "https://github.com/cioos-siooc/metadata-entry-form.git", subdirectory = "firebase_to_xml" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
